backend/requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.2
pydantic[email]
pyjwt==2.6.0
passlib[bcrypt]==1.7.4
setuptools
wheel
transformers
torch
python-multipart
mlflow
